requirements_windows.txt
akshare>=1.12.0
numpy>=1.24.0
pyautogui>=0.9.54
Pillow>=10.0.0
pytesseract>=0.3.10
pywin32>=306
